display bus destination rather than route name
The destination is what is actually displayed on the bus ticker, so this
change can reduce confusion. Also, Express buses are only reflected in
the destination name, not the route, so this means that users can now
see if the bus they're waiting for is an express or not.

To make the different routes clearer, colours have been added to the
route numbers. Next feature will be to add tooltips describing the route
name when the user taps on the route number.